% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daylength.R
\name{daylength}
\alias{daylength}
\title{daylength}
\usage{
daylength(sangle, LAT)
}
\arguments{
\item{LAT}{latitude at which the model is run}

\item{solar}{angle (in radians)}
}
\description{
This function calculates daylength based on solar angle and latitude as input. The Daylength Function is based on the BASFOR model by David Cameron and Marvel van Ojen.
}
\examples{
# get solar angle:
day<- seq(1,365*4,1)
sangle= -23.45*pi/180 * cos((pi/180) * (day +10)) #(radians)
plot(sangle)
dayl<- seq(1,365*4,1)
for (i in 1:length(dayl)){
 dayl[i]<-daylength(sangle[i],LAT=48.6542)
}
plot(dayl)
}
