% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DH_model_ins}
\alias{DH_model_ins}
\title{Example forcing data and parameters needed for running the DH model}
\format{
contains lists with data frames containing forcing data for two sites in France, alongside a parameter with best, upper and lower limits for the parameters
}
\source{
Data obtained from Cyrille Rathgeber,and used in http://www.plantphysiol.org/lookup/doi/10.1104/pp.16.00037, and modified to weekly resolution afterwards.
Cpool was obtained from Ogee et al Figure 3, and scaled for the purposes here. These are site data from locations where xylogenesis and wood anatomical data exists.
}
\usage{
DH_model_ins
}
\description{
Example forcing data and parameters needed for running the DH model
}
\examples{
data(DH_model_ins)
}
\keyword{datasets}
